Components/requirements.txt
wheel
tinyaes
pyinstaller
pyaesm
urllib3zxuqeqmf